@@ -1,23 +1,19 @@
-import React from 'react'
-import { Container } from "@material-ui/core"
-import { BrowserRouter, Switch, Route } from 'react-router-dom'
-import Home from './components/Home/Home.js'
-import { NavBar } from './components/NavBar/NavBar'
-import Auth from './components/Auth/Auth.js'
+import React from 'react';
+import { Container } from "@material-ui/core";
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Home from './components/Home/Home.js';
+import { NavBar } from './components/NavBar/NavBar';
+import Auth from './components/Auth/Auth.js';
 
 function App() {
-
-
   return (
-
     <BrowserRouter>
       <Container maxWidth="lg">
         <NavBar />
-        <Switch>
-          <Route path="/" exact component={Home} />
-          <Route path="/auth" exact component={Auth} />
-        </Switch>
-
+        <Routes>
+          <Route path="/" exact element={<Home />} />
+          <Route path="/auth" exact element={<Auth />} />
+        </Routes>
       </Container>
     </BrowserRouter>
   );
