@@ -1,42 +1,41 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect } from 'react';
 import { AppBar, Avatar, Button, Toolbar, Typography } from '@material-ui/core';
 import useStyles from './styles.js';
 import Instaverse from '../../images/Instaverse.png';
-import { Link, useHistory, useLocation } from 'react-router-dom';
-import { useDispatch } from 'react-redux'
-import decode from 'jwt-decode'
+import { Link, useNavigate, useLocation } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
+import { jwtDecode } from 'jwt-decode';
 
 export const NavBar = () => {
-
-    const classes = useStyles()
-    const [user, setUser] = useState(JSON.parse(localStorage.getItem('profile')))
-    const dispatch = useDispatch()
-    const history = useHistory()
-    const location = useLocation()
+    const classes = useStyles();
+    const [user, setUser] = useState(JSON.parse(localStorage.getItem('profile')));
+    const dispatch = useDispatch();
+    const navigate = useNavigate();
+    const location = useLocation();
 
     const logout = () => {
-        dispatch({ type: 'LOGOUT' })
+        dispatch({ type: 'LOGOUT' });
 
-        history.push('/auth')
+        navigate('/auth');
 
-        setUser(null)
-    }
+        setUser(null);
+    };
 
     useEffect(() => {
-        const token = user?.token
+        const token = user?.token;
 
-if(token) {
-    const decodedToken = decode(token)
+        if (token) {
+            const decodedToken = jwtDecode(token);
 
-    if(decodedToken.exp * 1000 < new Date().getTime()) logout()
-}
-        setUser(JSON.parse(localStorage.getItem('profile')))
-    }, [location, user?.token])
+            if (decodedToken.exp * 1000 < new Date().getTime()) logout();
+        }
+        setUser(JSON.parse(localStorage.getItem('profile')));
+    }, [location, user?.token]);
 
     return (
         <AppBar className={classes.appBar} position="static" color="inherit">
             <div className={classes.brandContainer}>
-                <Typography className={classes.heading} component={Link} to="/" variant="h2" align="center" >instaverse</Typography>
+                <Typography className={classes.heading} component={Link} to="/" variant="h2" align="center">instaverse</Typography>
                 <img className={classes.image} src={Instaverse} alt="instaverse" heigh="60" />
             </div>
             <Toolbar className={classes.toolbar}>
@@ -47,12 +46,9 @@ if(token) {
                         <Button variant="contained" className={classes.logout} onClick={logout} color="secondary">Logout</Button>
                     </div>
                 ) : (
-                    <Button component={Link} to="/auth" variant="contained" color="primary" >Sign In</Button>
-                )
-
-                }
+                    <Button component={Link} to="/auth" variant="contained" color="primary">Sign In</Button>
+                )}
             </Toolbar>
         </AppBar>
-    )
-}
-
+    );
+};
