@@ -1,24 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux'
-import { createStore, applyMiddleware, compose } from 'redux'
-import thunk from 'redux-thunk'
-import './index.css'
+import ReactDOM from 'react-dom/client';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware, compose } from 'redux';
+import thunk from 'redux-thunk';
+import './index.css';
 
 import App from './App';
 
-import reducers from './reducers/index.js'
+import reducers from './reducers/index.js';
 
-const store = createStore(reducers, compose(applyMiddleware(thunk)))
+const store = createStore(reducers, compose(applyMiddleware(thunk)));
 
-ReactDOM.render(
-  
-
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
     <Provider store={store}>
-      <App />
-    </Provider>, 
-    document.getElementById('root')
+        <App />
+    </Provider>
 );
-
-
-
